flask-server/requirements.txt
flask
flask_cors
openai
python-dotenv
requests
